Subject:  Added UI Elements from Sketch Design to Interface
Added an input accessory view to enable a user to set the mood upon an entry.  Added a custom UITableViewCell with correct assets for all devices.  Added maintenance code for UI functionality.  Need to update unit tests.

+++ FitJourney/DiaryEntry+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension DiaryEntry {
 
     @NSManaged var body: String
     @NSManaged var date: NSTimeInterval
-    @NSManaged var mood: Int16
+    @NSManaged var mood: Int
     @NSManaged var imageData: NSData?
     @NSManaged var location: String?
 }
